[task] Initial implementation of operations counting.

+++ include/StoreData.h
@@ -9,7 +9,7 @@ class StoreData
 {
     public:
             static void initializeFile(ofstream& myfile, vector<string> names);
-            static void appendData(ofstream& myfile, vector<double> data, int numVal);
+            static void appendData(ofstream& myfile, vector<long unsigned int> data, int numVal);
             static void closeFile(ofstream& myfile);
 };
